+++ jablib/src/test/java/org/jabref/logic/importer/ImportFormatReaderIntegrationTest.java
@@ -10,6 +10,7 @@ import java.util.stream.Stream;
 
 import javafx.collections.FXCollections;
 
+import org.jabref.logic.citationkeypattern.CitationKeyGenerator;
 import org.jabref.logic.citationkeypattern.CitationKeyPatternPreferences;
 import org.jabref.model.util.DummyFileUpdateMonitor;
 
@@ -30,10 +31,12 @@ class ImportFormatReaderIntegrationTest {
     void setUp() {
         ImporterPreferences importerPreferences = mock(ImporterPreferences.class, Answers.RETURNS_DEEP_STUBS);
         when(importerPreferences.getCustomImporters()).thenReturn(FXCollections.emptyObservableSet());
+        CitationKeyPatternPreferences citationKeyPatternPreferences = mock(CitationKeyPatternPreferences.class);
+        when(citationKeyPatternPreferences.getUnwantedCharacters()).thenReturn(CitationKeyGenerator.DEFAULT_UNWANTED_CHARACTERS);
         reader = new ImportFormatReader(
                 importerPreferences,
                 mock(ImportFormatPreferences.class, Answers.RETURNS_DEEP_STUBS),
-                mock(CitationKeyPatternPreferences.class), new DummyFileUpdateMonitor()
+                citationKeyPatternPreferences, new DummyFileUpdateMonitor()
         );
     }